Improve performance on country lookup
By using Angular's filter filter, when the text input changed, countries were being added and removed from the DOM and there was too much overhead -- it was slow. Now the elements are just hidden instead of removed and it is faster. Change some CSS as well to make the threes grid work fine with this setup.